% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cg_custom}
\alias{cg_custom}
\title{Conjugate Gradient for Large VI Problems}
\usage{
cg_custom(
  X,
  Z,
  P,
  omega,
  ridge_Z,
  ridge_X,
  s,
  offset_ridge_X,
  old_alpha,
  tol,
  it_max = 0L,
  low_dimension = 5L
)
}
\description{
Conjugate Gradient for Large VI Problems
}
