% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chol_calculate_outer_alpha}
\alias{chol_calculate_outer_alpha}
\title{Fast Calculation of E[alpha alpha^T] - Chol}
\usage{
chol_calculate_outer_alpha(cholL, P, alpha_mu, re_mu_list, re_group_list,
  re_position_list)
}
\description{
Fast Calculation of E[alpha alpha^T] - Chol
}
