% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{custom_glmer_samples}
\alias{custom_glmer_samples}
\title{Get samples from GLMER}
\usage{
custom_glmer_samples(glmer, samples, ordering)
}
\arguments{
\item{glmer}{object fitted using glmer}

\item{samples}{number of samples to draw}

\item{ordering}{order of output}
}
\description{
Order samples from glmer to match names from vglmer.
}
\keyword{internal}
