% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmer_regression.R
\name{custom_HMC_samples}
\alias{custom_HMC_samples}
\title{Get HMC samples but ordered to match vector}
\usage{
custom_HMC_samples(HMC, ordering)
}
\description{
Get HMC samples but ordered to match vector
}
