% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{ELBO}
\alias{ELBO}
\title{Get ELBO from fitted variational model}
\usage{
ELBO(x, type = "final")
}
\arguments{
\item{type}{Default is "final" giving ELBO at convergence. "trajectory" gives
full sequence of ELBOs at each iteration.}
}
\description{
Get ELBO from fitted variational model
}
