% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmer_regression.R
\name{EM_prelim}
\alias{EM_prelim}
\title{Simple EM algorithm for starting values.}
\usage{
EM_prelim(X, Z, s, pg_b, iter, ridge = 2)
}
\description{
Use ridge penalty to prevent separation. Not be called by user!
}
