% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_functions.R
\name{fallback_interpret.gam0}
\alias{fallback_interpret.gam0}
\title{Interpret a vglmer formula for splines}
\usage{
fallback_interpret.gam0(gf, textra = NULL, extra.special = NULL)
}
\arguments{
\item{gf}{A vglmer formula}

\item{textra}{Unused internal argument}

\item{extra.special}{Allow extra special terms to be passed}
}
\description{
A modified version of interpret.gam0 from mgcv. Used when mgcv's
  interpret.gam fails; usually when some environment object is passed to v_s.
}
\keyword{internal}
