% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_functions.R
\name{fallback_subbars}
\alias{fallback_subbars}
\title{To be used if subbars fails, usually when there is an argument to 
v_s ; adapted from lme4}
\usage{
fallback_subbars(term)
}
\arguments{
\item{term}{a formula with lme4-style syntax; see \link[lme4]{subbars}}
}
\description{
To be used if subbars fails, usually when there is an argument to 
v_s ; adapted from lme4
}
\keyword{internal}
