% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{format_obj}
\alias{format_obj}
\alias{format_glmer}
\alias{format_vglmer}
\alias{format_stan}
\title{Format Existing Objects}
\usage{
format_glmer(object)

format_vglmer(object)

format_stan(object, useSigma = FALSE)
}
\arguments{
\item{object}{Object from glmer, vglmer, or stan to the respective "format_"
function}

\item{useSigma}{Return variance component parameters from STAN? Default "FALSE".}
}
\description{
Takes a regression output from glmer, stan, or vglmer and extracts the fixed
and random effects.
}
