% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_functions.R
\name{formOmega}
\alias{formOmega}
\title{Code from Wand and Ormerod (2008)
Found here here: 10.1111/j.1467-842X.2008.00507.x}
\usage{
formOmega(a, b, intKnots)
}
\arguments{
\item{a}{lower boundary}

\item{b}{upper boundary}

\item{intKnots}{internal knots}
}
\description{
Code from Wand and Ormerod (2008)
Found here here: 10.1111/j.1467-842X.2008.00507.x
}
