% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{custom_HMC_linpred}
\alias{custom_HMC_linpred}
\alias{hmc_samples}
\alias{custom_HMC_samples}
\title{Get HMC samples but ordered to match vector of names provided}
\usage{
custom_HMC_linpred(HMC, data)

custom_HMC_samples(HMC, ordering)
}
\arguments{
\item{HMC}{Object from rstanarm}

\item{data}{Data to get predictions on.}

\item{ordering}{vector of names to order the posterior samples.}
}
\description{
Get HMC samples but ordered to match vector of names provided
}
\keyword{internal}
