% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LinRegChol}
\alias{LinRegChol}
\title{Linear Regression by Cholesky}
\usage{
LinRegChol(X, omega, prior_precision, y, adj_y, save_chol = TRUE)
}
\arguments{
\item{X}{Design Matrix}

\item{omega}{Polya-Gamma weights}

\item{prior_precision}{Prior Precision for Regression}

\item{y}{Outcome}

\item{save_chol}{Save cholesky factor}

\item{adjustment}{vector}
}
\description{
Do linear regression of form (X^T O X + P)^{-1} X^T y where O is omega, P is
precision.
}
\keyword{internal}
