% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{MAVB_linpred}
\alias{MAVB_linpred}
\title{Linear Predictor Following MAVB}
\usage{
MAVB_linpred(model, newdata, samples, var_px = Inf, summary = TRUE)
}
\arguments{
\item{model}{Model fit using vglmer}

\item{newdata}{Data to get predictions on.}

\item{samples}{Samples to draw from MAVB distribution.}

\item{var_px}{Default (Inf); variance of working prior. Higher is more
diffuse and thus likely better.}

\item{summary}{Return summary of linear predictor, not full posterior.}
}
\description{
Combine prediction and MAVB into a single function.
}
