% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_functions.R
\name{Predict.matrix.randwalk.smooth}
\alias{Predict.matrix.randwalk.smooth}
\title{Predict Methods for random walk smooth}
\usage{
\method{Predict.matrix}{randwalk.smooth}(object, data)
}
\arguments{
\item{object}{a smooth object; see documentation for other methods in
\code{mgcv}.}

\item{data}{a data.frame; see documentation for other methods in
\code{mgcv}.}
}
\description{
Predict Methods for random walk smooth
}
\keyword{internal}
