% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fixef}
\alias{ranef}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lme4}{\code{\link[lme4]{fixef}}, \code{\link[lme4]{ranef}}}
}}

