% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{simple_EM}
\alias{simple_EM}
\alias{EM_prelim_logit}
\alias{EM_prelim_nb}
\title{Simple EM algorithm for starting values.}
\usage{
EM_prelim_logit(X, Z, s, pg_b, iter, Z_list, Tinv, cyclical_pos, ridge = 2)

EM_prelim_nb(X, Z, y, est_r, iter, ridge = 2)
}
\arguments{
\item{X}{Design matrix}

\item{Z}{RE design matrix}

\item{s}{(y_i - n_i)/2 for polya-gamma input}

\item{pg_b}{n_i as vector input}

\item{iter}{iterations}

\item{ridge}{variance of ridge prior}

\item{y}{Raw observed y_i}

\item{est_r}{Initial r value (not updated!)}
}
\description{
Use ridge penalty to prevent separation. Not be called by user!
}
\keyword{internal}
