% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_functions.R
\name{smooth.construct.randwalk.smooth.spec}
\alias{smooth.construct.randwalk.smooth.spec}
\title{Constructor for random walk smooth}
\usage{
\method{smooth.construct}{randwalk.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth object; see documentation for other methods in
\code{mgcv}.}

\item{data}{a data.frame; see documentation for other methods in
\code{mgcv}.}

\item{knots}{not used}
}
\description{
Constructor for random walk smooth
}
\details{
See the function \link{v_s} for details.
}
\keyword{internal}
