% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_functions.R
\name{v_fe}
\alias{v_fe}
\title{Fixed Effects}
\usage{
v_fe(group, interactions = ~1)
}
\description{
This estimates (high-dimensional) fixed effects using the results in Auslen
and Goplerud (2025). Its syntax is experimental but \code{y ~ v_fe(id) + (1 |
x)} would estimate a fixed effect for \code{"id"} and a random effect for \code{"x"}.
}
