% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_functions.R
\name{v_s}
\alias{v_s}
\title{Spline Creator}
\usage{
v_s(
  ...,
  type = "tpf",
  knots = NULL,
  by = NA,
  by_re = TRUE,
  add_linear = TRUE,
  outer_okay = FALSE
)
}
\arguments{
\item{...}{variable name}

\item{type}{Spline basis: "tpf" for truncated linear splines or "o" for
O-Sullivan (see Wand and Ormerod 2008)}

\item{knots}{Provide either number of knots (evenly spaced at quantiles), a
grid of knots. Default ("NULL") uses K = min(N/4, 35) where "N" is number
of unique values.}

\item{by}{Interact the spline with some covariate.}

\item{by_re}{If interacted with a factor variable, should the "base"
interactions be regularized? Default is "TRUE".}

\item{outer_okay}{Can values in "x" exceed the knots? Default "FALSE".}
}
\description{
A simple function to estimate splines, similar to "s" in "mgcv". Allows for
truncated (linear) splines or O'Sullivan splines.
}
\references{
Wand, Matt P. and Ormerod, John T. 2008. On semiparametric regression with
O'Sullivan penalized splines. Australian & New Zealand Journal of Statistics.
50(2): 179-198.
}
