% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmer_regression.R
\name{var_mat}
\alias{var_mat}
\alias{rowVar}
\alias{colVar}
\title{Variance of Rows or Columns of Matrices}
\usage{
rowVar(matrix)

colVar(matrix)
}
\arguments{
\item{matrix}{Matrix of numeric inputs.}
}
\description{
Base R implementation for variance. Analogue of rowMeans.
}
