% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{vglmer-class}
\alias{vglmer-class}
\alias{fixef.vglmer}
\alias{sigma.vglmer}
\alias{ranef.vglmer}
\alias{coef.vglmer}
\alias{vcov.vglmer}
\alias{print.vglmer}
\alias{summary.vglmer}
\alias{formula.vglmer}
\alias{format_vglmer}
\alias{ELBO}
\title{Generic Functions after Running vglmer}
\usage{
\method{fixef}{vglmer}(object, ...)

\method{sigma}{vglmer}(object, ...)

\method{ranef}{vglmer}(object, ...)

\method{coef}{vglmer}(object, ...)

\method{vcov}{vglmer}(object, ...)

\method{print}{vglmer}(x, ...)

\method{summary}{vglmer}(object, display_re = TRUE, ...)

\method{formula}{vglmer}(x, form, ...)

format_vglmer(object)

ELBO(object, type = c("final", "trajectory"))
}
\arguments{
\item{object}{A model estimated using \code{vglmer}.}

\item{...}{Not used.}

\item{x}{Model fit using vglmer}

\item{display_re}{Logical. Default (\code{TRUE}) prints a summary of the
random effects alongside the fixed effects.}

\item{form}{Character. Describes the type of formula to report:
\code{"original"} returns the user input, \code{"fe"} returns the fixed
effects only, \code{"re"} returns the random effects only.}

\item{type}{Character. Default (\code{"final"}) gives the ELBO at convergence.
\code{"trajectory"} gives the ELBO estimated at each iteration. This is
used to assess model convergence.}
}
\description{
\code{vglmer} is designed to allow for standard methods from \code{lm} and
\code{lme4} to be used with limited changes on models estimated using
\code{vglmer}. These all provide summaries of the estimated posterior.
}
\details{
The accompanying functions are briefly described below. \code{coef}
  and \code{vcov} return the mean and variance of the fixed effects.
  \code{fixef} returns the mean of the fixed effects.

  \code{ranef} extracts the random effects in a similar, although slightly
  different format, to \code{lme4}. It includes the estimated posterior mean
  and variance in a list of data.frames with one entry per random effect.

  \code{format_vglmer} collects the mean and variance of the fixed and random
  effects into a single data.frame. This is useful for examining all of the
  posterior estimates simultaneously. See \code{?format_glmer} for a function
  that converts an object estimated with \code{[g]lmer} into a comparable
  format.
  
  \code{ELBO} extracts the ELBO from the estimated model. \code{type} can be
  set equal to \code{"trajectory"} to get the estimated ELBO at each
  iteration and assess convergence.
}
