% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{expand_formula_clogit}
\alias{expand_formula_clogit}
\alias{expand_data_clogit}
\alias{vglmer_clogit}
\alias{vglmer_OvR}
\title{Estimate with vglmer for categorical outcomes}
\usage{
expand_formula_clogit(formula, choice_names, exempt = NULL)

expand_data_clogit(data, response, id_name)

vglmer_clogit(
  formula,
  data,
  family,
  pooled = TRUE,
  id_name,
  choices,
  control = vglmer_control()
)

vglmer_OvR(formula, data, family, control = vglmer_control(), pooled = TRUE)
}
\description{
Estimate with vglmer for categorical outcomes
}
