% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_functions.R
\name{vglmer_build_fe}
\alias{vglmer_build_fe}
\title{Create a sparse design matrix for the fixed effects}
\usage{
vglmer_build_fe(group, inter, levels = NULL)
}
\description{
Create a sparse design matrix for the fixed effects
}
