% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_functions.R
\name{vglmer_interpret.gam0}
\alias{vglmer_interpret.gam0}
\title{A modified version of interpret.gam0 from mgcv.
Copied almost verbatim but adjusted to allow custom objects
to be parsed as arguments to "v_s"}
\usage{
vglmer_interpret.gam0(gf, textra = NULL, extra.special = NULL)
}
\description{
A modified version of interpret.gam0 from mgcv.
Copied almost verbatim but adjusted to allow custom objects
to be parsed as arguments to "v_s"
}
