% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmer_regression.R
\name{vglmer_logit}
\alias{vglmer_logit}
\title{Regression}
\usage{
vglmer_logit(formula, data, iterations, factorization.method,
  tolerance_elbo = 1e-04, prevent_degeneracy = FALSE, prior_variance,
  parameter.expansion = "mean", random.seed = 1, debug.ELBO = FALSE,
  print.prog = NULL, tolerance_parameters = 1e-04, quiet = T,
  init = "EM")
}
\description{
Regression
}
