% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_functions.R
\name{format_obj}
\alias{format_obj}
\alias{format_glmer}
\title{Format Existing Objects}
\usage{
format_glmer(object)
}
\arguments{
\item{object}{Object from glmer or lmer to the respective "format_"
function}
}
\description{
Takes a regression output from glmer, stan, or vglmer and extracts the fixed
and random effects.
}
\keyword{internal}
