% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{MAVB}
\alias{MAVB}
\title{Perform MAVB after fitting vglmer}
\usage{
MAVB(object, samples, verbose = FALSE, var_px = Inf)
}
\arguments{
\item{object}{Model fit using vglmer}

\item{samples}{Samples to draw from MAVB distribution.}

\item{verbose}{Show progress of MAVB.}

\item{var_px}{Default (Inf); variance of working prior. Higher is more
diffuse and thus likely better.}
}
\description{
Given a model from vglmer, perform marginally augmented variational Bayes
(MAVB) to improve approximation quality; Goplerud (2022) provides details. At
present, it is only enabled for binomial models.
}
\details{
This should only be used if the parameters of the model are of interest; to
use MAVB to improve prediction accuracy, use predict_MAVB.
}
\references{
Goplerud, Max. 2022. "Fast and Accurate Estimation of Non-Nested Binomial
Hierarchical Models Using Variational Inference." Bayesian Analysis. 17(2):
623-650.
}
