% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{MAVB}
\alias{MAVB}
\title{Perform MAVB after fitting vglmer}
\usage{
MAVB(object, samples, var_px = Inf)
}
\arguments{
\item{object}{Model fit using vglmer}

\item{samples}{Samples to draw from MAVB distribution.}

\item{var_px}{Default (Inf); variance of working prior. Higher is more
diffuse and thus likely better.}
}
\description{
Given a model from vglmer, perform MAVB to improve approximation quality. See
dissertation for details. This function uses a naive loop so is slower than
necessary. This will be updated shortly.
}
