% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{MAVB}
\alias{MAVB}
\title{Perform MAVB after fitting vglmer}
\usage{
MAVB(object, samples, verbose = FALSE, var_px = Inf)
}
\arguments{
\item{object}{Model fit using \code{vglmer}.}

\item{samples}{Number of samples to draw.}

\item{verbose}{Show progress of MAVB.}

\item{var_px}{Variance of working prior for marginal augmentation. Default
(\code{Inf}) is a flat, improper, prior.}
}
\value{
This returns a matrix with \code{sample} rows for each fixed and
  random effect: \eqn{q(\alpha,\beta)}.
}
\description{
Given a model from vglmer, this function performs marginally augmented
variational Bayes (MAVB) to improve approximation quality; Goplerud (2022)
provides details. At present, it is only enabled for binomial models. This
should only be used if the parameters of the model are of interest; to use
MAVB when generating predictions, use \link{predict_MAVB}.
}
\references{
Goplerud, Max. 2022. "Fast and Accurate Estimation of Non-Nested Binomial
Hierarchical Models Using Variational Inference." Bayesian Analysis. 17(2):
623-650.
}
