% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{posterior_samples.vglmer}
\alias{posterior_samples.vglmer}
\title{Draw samples from the variational distribution}
\usage{
posterior_samples.vglmer(object, samples, verbose = FALSE)
}
\arguments{
\item{object}{Model fit using \code{vglmer}.}

\item{samples}{Number of samples to draw.}

\item{verbose}{Show progress in drawing samples.}
}
\description{
This function draws samples from the estimated variational
  distributions. If using \code{MAVB} to improve the quality of the
  approximating distribution, please use \link{MAVB} or \link{predict_MAVB}.
}
