% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R
\name{posterior_samples.vglmer}
\alias{posterior_samples.vglmer}
\title{Draw samples from the (non-MAVB) posterior}
\usage{
posterior_samples.vglmer(object, samples, verbose = FALSE)
}
\arguments{
\item{object}{Model fit using vglmer}

\item{samples}{Samples to draw from MAVB distribution.}

\item{verbose}{Print the progress of the samples.}
}
\description{
Draw samples from the (non-MAVB) posterior
}
