% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_functions.R
\name{predict.vglmer_multiclass}
\alias{predict.vglmer_multiclass}
\title{Multiclass Prediction}
\usage{
\method{predict}{vglmer_multiclass}(object, newdata, allow_missing_levels = FALSE)
}
\description{
Multiclass Prediction
}
