% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{summary_vglmer}
\alias{summary_vglmer}
\alias{fixef.vglmer}
\alias{coef.vglmer}
\alias{vcov.vglmer}
\title{Summarize output from vglmer}
\usage{
\method{fixef}{vglmer}(object, ...)

\method{coef}{vglmer}(object, ...)

\method{vcov}{vglmer}(object, ...)
}
\arguments{
\item{object}{Model fitted from vglmer}
}
\description{
Allows the use of standard methods from lm or lmer to sumarize the posterior
output. The format for ranef differs from lmer and is shown below.
}
\details{
\code{coef} and \code{vcov} return the mean and variance of the fixed effects. \code{fixef} is a
synonym for \code{coef}.
}
