% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{vglmer-class}
\alias{vglmer-class}
\alias{fixef.vglmer}
\alias{sigma.vglmer}
\alias{ranef.vglmer}
\alias{coef.vglmer}
\alias{vcov.vglmer}
\alias{print.vglmer}
\alias{summary.vglmer}
\alias{formula.vglmer}
\alias{format_vglmer}
\alias{ELBO}
\title{Generic Functions after Running vglmer}
\usage{
\method{fixef}{vglmer}(object, ...)

\method{sigma}{vglmer}(object, ...)

\method{ranef}{vglmer}(object, ...)

\method{coef}{vglmer}(object, ...)

\method{vcov}{vglmer}(object, ...)

\method{print}{vglmer}(x, ...)

\method{summary}{vglmer}(object, display_re = TRUE, ...)

\method{formula}{vglmer}(x, form, ...)

format_vglmer(object)

ELBO(object, type = c("final", "trajectory"))
}
\arguments{
\item{object}{Object from vglmer}

\item{...}{Not used.}

\item{x}{Model fit using vglmer}

\item{display_re}{Print summary of random effects. Default is TRUE}

\item{form}{What type of formula to return? "original" is user provided; "fe"
is fixed effect only; "re" is random effect only.}

\item{type}{Default is "final" giving ELBO at convergence. "trajectory" gives
full sequence of ELBOs at each iteration.}
}
\value{
\code{coef} and \code{vcov} return the mean and variance of the fixed
effects. \code{fixef} is a synonym for \code{coef}. \code{ranef} extracts the
random effects in a similar, although slightly different, format to
\code{lme4}.

\code{format_vglmer} collects the mean and variance of the fixed and random
effects into a data.frame
}
\description{
Allows the use of standard methods from lm or lmer to sumarize the posterior
output.
}
