% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmer_regression.R
\name{vglmer}
\alias{vglmer}
\title{Variational Inference for Non-Linear Hierarchical Models}
\usage{
vglmer(formula, data, family, control = vglmer_control())
}
\arguments{
\item{formula}{Standard glmer-style formula for random effects.}

\item{data}{data.frame containing the outcome and variables.}

\item{family}{Options are "binomial" or "negbin". If "binomial", outcome must
be either {0,1} (binary) or cbind(success, failure) as per standard glm(er)
syntax. Non-integer values are permitted for binomial if "force_whole" is
set to FALSE in vglmer_control.}

\item{control}{Control additional arguments. Must be made using
vglmer_control(); see for documentation for additional details.}
}
\description{
Estimate a hierarchical modelusing mean-field variational
inference. Accepts standard syntax to glmer: y ~ X + (1 + Z | g). Options are
described below.
}
