% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmer_regression.R
\name{vglmer_control}
\alias{vglmer_control}
\title{Control for vglmer}
\usage{
vglmer_control(
  iterations = 1000,
  collapse_size = "FE",
  block_collapse = FALSE,
  prior_variance = "mean_exists",
  factorization_method = "weak",
  tolerance_elbo = 1e-08,
  tolerance_parameters = 1e-05,
  prevent_degeneracy = FALSE,
  force_whole = TRUE,
  verbose_time = TRUE,
  parameter_expansion = "mean",
  random_seed = 1,
  do_timing = FALSE,
  debug_param = FALSE,
  return_data = FALSE,
  linpred_method = "joint",
  vi_r_method = "VEM",
  vi_r_val = NA,
  do_SQUAREM = TRUE,
  verify_columns = FALSE,
  debug_ELBO = FALSE,
  print_prog = NULL,
  quiet = T,
  quiet_rho = TRUE,
  init = "EM_FE"
)
}
\arguments{
\item{iterations}{Number of iterations for the model.}

\item{prior_variance}{Options are jeffreys, mcmcglmm, mvD, mean_exists,
limit, and uniform. They are defined as an Inverse Wishart with the
following parameters where d is the dimensionality of the random effect. At
present, all models use "mean_exists" for a well-defined proper prior.
\itemize{
\item kn: See Kass and Natarajan (2006)
\item jeffreys: IW(0, 0)
\item mcmcglmm: IW(0, I)
\item mvD: IW(-d, I)
\item mean_exists: IW(d + 1, I)
\item limit: IW(d - 1, 0)
\item uniform: IW(-[d+1], 0)
}
The model may fail to converge if an improper prior is used.}

\item{factorization_method}{The factorization method to use. Described in
detail in the dissertation. strong, partial, and weak correspond to Schemes
I, II, and III respectively. "weak" should have best performance but is
slowest.}

\item{tolerance_elbo}{Change in ELBO to stop algorithm.}

\item{tolerance_parameters}{Change in value of any parameter to stop algorithm.}

\item{prevent_degeneracy}{Ignored for the moment.}

\item{force_whole}{Require whole numbers. Set to FALSE to allow "quasi-binomial".}

\item{verbose_time}{Print time for each step (debugging only)}

\item{parameter_expansion}{At moment, accepts 'mean' or 'none'. 'mean' is
costless and should always be used!}

\item{random_seed}{Set seed for initialization.}

\item{do_timing}{Estimate timing with tictoc}

\item{debug_param}{Debug parameter convergence.}

\item{return_data}{Return the design (X,Z) for debugging afterwards.}

\item{linpred_method}{Method for updating means of beta and alpha. "joint" is best.}

\item{vi_r_method}{Type of estimate for "r"; at moment, "fixed" (provide r),
"VEM" (treat r as point estimate; default);
"Laplace" (estimate using Laplace approximation described in the Addendum on GitHub); or "delta"
(experimential).}

\item{vi_r_val}{For fixed "r", which value?}

\item{do_SQUAREM}{Accelerate method using SQUAREM}

\item{debug_ELBO}{Debug ELBO trajectory.}

\item{print_prog}{Print after print_prog iterations to show progress.}

\item{quiet}{Don't print noisy intermediate output.}

\item{init}{Initialization method can be EM, zero, or random at moment.}
}
\description{
Provides a set of control arguments to vglmer
}
