% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{vglmer_predict}
\alias{vglmer_predict}
\alias{predict.vglmer}
\title{Predict after vglmer}
\usage{
\method{predict}{vglmer}(
  object,
  newdata,
  samples = 0,
  samples_only = FALSE,
  summary = TRUE,
  allow_missing_levels = FALSE,
  ...
)
}
\arguments{
\item{object}{Object from vglmer.}

\item{newdata}{Data to get predictions on.}

\item{samples}{How many samples to draw? 0, default, gets the expected value.
Two methods: 
\itemize{
\item Number - draw that many samples.
\item Matrix - use previous samples for prediction. Currently used for MAVB
but will be depreciated in updated version.
}}

\item{samples_only}{Return only samples *not* linear predictor.}

\item{summary}{Return summary of linear predictor, not full posterior.}

\item{allow_missing_levels}{Allow prediction for random effects not in object.
As is standard, give an estimate of "0" for that effect.}

\item{...}{Not used; included to maintain compatability}
}
\description{
Get linear predictor for new observations after using vglmer.
}
\examples{

sim_data <- data.frame(x = rnorm(100), 
y = rbinom(100, 1, 0.5), 
g =sample(letters, 100, replace = TRUE))

# Run with defaults
est_vglmer <- vglmer(y ~ x + (x|g), data = sim_data, family = "binomial")

# Simple prediction 
predict(est_vglmer, newdata = sim_data)

\dontrun{
#Fails!
predict(est_vglmer, newdata = data.frame(g = "AB", x = 0))
}
predict(est_vglmer, newdata = data.frame(g = "AB", x = 0), 
allow_missing_levels = TRUE)
}
