% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmer_regression.R
\name{vglmer_predict}
\alias{vglmer_predict}
\title{Predict linear predictor vglmer.}
\usage{
vglmer_predict(model, newdata, samples, samples.only = FALSE,
  summary = TRUE)
}
\description{
Predict linear predictor vglmer.
}
