% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAVB_functions.R, R/predict_functions.R
\name{predict_MAVB}
\alias{predict_MAVB}
\alias{vglmer_predict}
\alias{predict.vglmer}
\title{Predict after vglmer}
\usage{
predict_MAVB(
  object,
  newdata,
  samples = 0,
  samples_only = FALSE,
  var_px = Inf,
  summary = TRUE,
  allow_missing_levels = FALSE
)

\method{predict}{vglmer}(
  object,
  newdata,
  samples = 0,
  samples_only = FALSE,
  summary = TRUE,
  allow_missing_levels = FALSE,
  ...
)
}
\arguments{
\item{object}{Object from vglmer.}

\item{newdata}{Data to get predictions on.}

\item{samples}{How many samples to draw? 0, default, gets the expected value.
Two methods:
\itemize{
\item Number - draw that many samples.
\item Matrix - use previous samples for prediction. Currently used for MAVB
but will be depreciated in updated version.
}}

\item{samples_only}{Return only samples *not* linear predictor.}

\item{var_px}{Default (Inf); variance of working prior. Higher is more
diffuse and thus likely better.}

\item{summary}{Return summary of linear predictor, not full posterior.}

\item{allow_missing_levels}{Allow prediction for random effects not in object.
As is standard, give an estimate of "0" for that effect.}

\item{...}{Not used; included to maintain compatability}
}
\value{
Returns an estimate of the linear predictor. The default returns the
  predicted posterior mean. If "samples > 0", then it returns a summary of
  the prediction for each observation, i.e. its mean and variance. Setting "summary = FALSE" will return \code{samples} posterior samples of
  the linear predictor for each observation.

  \code{predict_MAVB} performs MAVB as described in Goplerud (2020) and then
  returns either a posterior summary or the samples using the same options as
  the generic predict.
  If "allow_missing_levels = TRUE", then observations with a new
  (unseen) level for the random effect get a "zero" for that term of the
  prediction.
}
\description{
Get linear predictor for new observations after using vglmer. Use predict_MAVB
to combine linear prediction after performing MAVB.
}
\examples{

sim_data <- data.frame(
  x = rnorm(100),
  y = rbinom(100, 1, 0.5),
  g = sample(letters, 100, replace = TRUE)
)

# Run with defaults
est_vglmer <- vglmer(y ~ x + (x | g), data = sim_data, family = "binomial")

# Simple prediction
predict(est_vglmer, newdata = sim_data)
# Return 10 posterior draws of the linear predictor for each observation.
predict_MAVB(est_vglmer, newdata = sim_data, summary = FALSE, samples = 10)
\dontrun{
# Fails!
predict(est_vglmer, newdata = data.frame(g = "AB", x = 0))
}
# Works
predict(est_vglmer,
  newdata = data.frame(g = "AB", x = 0),
  allow_missing_levels = TRUE
)
}
